Refactoring -- export termux-service module
