+++ app/src/main/java/com/termux/app/TermuxService.java
@@ -22,8 +22,10 @@ import android.util.Log;
 import android.widget.ArrayAdapter;
 
 import com.termux.R;
+import com.termux.service.BackgroundJob;
 import com.termux.service.TermuxConfig;
 import com.termux.service.TermuxEnvironment;
+import com.termux.service.TermuxInstaller;
 import com.termux.terminal.EmulatorDebug;
 import com.termux.terminal.TerminalSession;
 import com.termux.terminal.TerminalSession.SessionChangedCallback;
