+++ terminal-service/src/main/cpp/Android.mk
