+++ terminal-service/src/main/cpp/termux-bootstrap.c
@@ -1,0 +1,11 @@
+#include <jni.h>
+
+extern jbyte blob[];
+extern int blob_size;
+
+JNIEXPORT jbyteArray JNICALL Java_com_termux_service_TermuxInstaller_getZip(JNIEnv *env,
+                                                                            __attribute__((__unused__)) jobject This) {
+    jbyteArray ret = (*env)->NewByteArray(env, blob_size);
+    (*env)->SetByteArrayRegion(env, ret, 0, blob_size, blob);
+    return ret;
+}
