+++ terminal-service/src/main/java/com/termux/service/TermuxEnvironment.java
